Add files via upload
This Python program receives 3 .txt files, each file representing a count of 3 boxes size: Large: 3 by 3 by 3, Medium: 2 by 2 by 2 and Small: 1 by 1 by 1 by 1. 
Every line in each file represents a box.
The program maps the space of a trailer and arranges the boxes in the 3D space, allowing the user of the program to know where the boxes go and if they fit within the trailer. 

As a quick example of the complexity of this problem, a Large box of 3 x 3 x 3 has a total of 27 cubic units, and a Medium box of 2 x 2 x 2 is only 8 cubic units, so volumetrically speaking, in theory, you could fit the volume of 3 Medium boxes (2x2x2) in the space of 1 Large box (3x3x3), but in reality, to make that work you will need to cut the boxes to make that work. Respecting the integrity of the boxes only 1 Medium box (2x2x2) fits within the space of a Large one (3x3x3), and it also allows for a maximum of 19 Small boxes of 1x1x1.

As a second challenge, this program also receives a .txt file with points in a cartesian plane, and it calculates the shortest route among all the points.